+++ src/TallaEgg.Core/Enums/Order/SymbolStatus.cs
@@ -1,0 +1,9 @@
+namespace TallaEgg.Core.Enums.Order
+{
+    public enum SymbolStatus
+    {
+        Active,
+        Inactive,
+        Suspended
+    }
+}